refactor: 機械的クリーンアップとデッドコード除去
clippy 全件を解消し、未使用の型/関数/export を削除（挙動不変・純減）

+++ crates/pika-cli/src/main.rs
@@ -421,7 +421,7 @@ mod win {
         unsafe {
             SHChangeNotify(
                 SHCNE_ASSOCCHANGED as i32,
-                SHCNF_IDLIST as u32,
+                SHCNF_IDLIST,
                 ptr::null(),
                 ptr::null(),
             );
